refactor: Restructure test functions for better organization
Restructured test functions to improve clarity and maintainability, and
to reflect the new modular project structure.

- Created separate test files for different functionalities:
  - Moved common utility tests to tests/test_file_utils.py and
    tests/test_web_utils.py
  - Kept PDF scraper tests in tests/test_pdf_scraper.py
- Updated imports in test files to match the new project structure
- Added mock patches to ensure isolation of tests and to handle external
  dependencies

# Issue References:
# - Related to #4
